fix ios nonce type

+++ plugins/nkn-sdk-flutter/ios/Classes/impl/Wallet.swift
@@ -201,7 +201,7 @@ class Wallet : ChannelBase, IChannelHandler, FlutterStreamHandler {
         let address = args["address"] as? String
         let amount = args["amount"] as? String
         let fee = args["fee"] as! String
-        let nonce = args["nonce"] as? Int64
+        let nonce = args["nonce"] as? Int
         let attributes = args["attributes"] as? FlutterStandardTypedData
         let seedRpc = args["seedRpc"] as? [String]
         
@@ -231,7 +231,7 @@ class Wallet : ChannelBase, IChannelHandler, FlutterStreamHandler {
             let transactionConfig: NknTransactionConfig = NknTransactionConfig()
             transactionConfig.fee = fee
             if (nonce != nil) {
-                transactionConfig.nonce = nonce!
+                transactionConfig.nonce = Int64(nonce!)
             }
             if(attributes != nil){
                 transactionConfig.attributes = attributes?.data
